Adding folder structure & Kotlin files
