+++ app/src/main/java/com/example/cinemates/adapter/ItemsRecyclerViewAdapter.kt
@@ -6,9 +6,7 @@ import androidx.recyclerview.widget.RecyclerView
 import android.view.ViewGroup
 import android.view.LayoutInflater
 import com.example.cinemates.model.data.Cast
-import android.content.Intent
 import android.view.View
-import com.example.cinemates.view.ui.ActorDetailsFragment
 import androidx.recyclerview.widget.DiffUtil
 import com.example.cinemates.NavGraphDirections
 import com.example.cinemates.databinding.ListItemMovieLongBinding
