+++ app/src/main/java/com/example/cinemates/view/ui/home/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.cinemates.view.ui;
+package com.example.cinemates.view.ui.home;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
