+++ app/src/main/java/com/example/cinemates/view/ui/profile/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.example.cinemates.view.ui;
+package com.example.cinemates.view.ui.profile;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
