+++ app/src/main/java/com/example/cinemates/view/ui/profile/ProfileViewModel.kt
@@ -1,0 +1,10 @@
+package com.example.cinemates.view.ui.profile
+
+import androidx.lifecycle.ViewModel
+
+/**
+ * @author Antonio Di Nuzzo
+ * Created 24/08/2022
+ */
+class ProfileViewModel : ViewModel() {
+}
