+++ app/src/main/java/com/example/cinemates/view/ui/saved/SavedFragment.java
@@ -1,4 +1,4 @@
-package com.example.cinemates.view.ui;
+package com.example.cinemates.view.ui.saved;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
