+++ app/src/main/java/com/example/cinemates/view/ui/saved/SavedViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.cinemates.view.ui.saved
+
+/**
+ * @author Antonio Di Nuzzo
+ * Created 24/08/2022
+ */
+class SavedViewModel {
+}
