+++ app/src/main/java/com/example/cinemates/view/ui/search/SearchMovieFragment.java
@@ -1,4 +1,4 @@
-package com.example.cinemates.view.ui;
+package com.example.cinemates.view.ui.search;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,38 +13,40 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.cinemates.adapter.ItemsRecyclerViewAdapter;
-import com.example.cinemates.databinding.FragmentSearchActorBinding;
+import com.example.cinemates.databinding.FragmentSearchMovieBinding;
 import com.example.cinemates.interfaces.CustomizableFragment;
-import com.example.cinemates.model.data.Cast;
+import com.example.cinemates.model.data.Movie;
 import com.example.cinemates.util.ViewSize;
 import com.example.cinemates.view.viewmodel.MovieViewModel;
 
 import java.util.List;
 
+public class SearchMovieFragment extends Fragment implements CustomizableFragment {
 
-public class SearchActorFragment extends Fragment implements CustomizableFragment {
-
-    private FragmentSearchActorBinding mBinding;
-    private ItemsRecyclerViewAdapter<Cast> mRecyclerViewAdapter;
+    private FragmentSearchMovieBinding mBinding;
+//    private MovieRecyclerViewAdapter mRecyclerViewAdapter;
+private ItemsRecyclerViewAdapter<Movie> mAdapter;
     private MovieViewModel mViewModel;
-
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mRecyclerViewAdapter = new ItemsRecyclerViewAdapter<>(ViewSize.LONG);
+        mAdapter = new ItemsRecyclerViewAdapter<>(ViewSize.SMALL);
+
         mViewModel = new ViewModelProvider(getActivity()).get(MovieViewModel.class);
 
-
     }
+
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        mBinding = FragmentSearchActorBinding.inflate(inflater, container, false);
-        mBinding.recyclerView.setAdapter(mRecyclerViewAdapter);
+        mBinding = FragmentSearchMovieBinding.inflate(inflater, container, false);
+
+        mBinding.recyclerView.setAdapter(mAdapter);
         mBinding.recyclerView.setEmptyView(mBinding.emptyView.getRoot());
+
 
         return mBinding.getRoot();
     }
@@ -52,10 +54,11 @@ public class SearchActorFragment extends Fragment implements CustomizableFragmen
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        mViewModel.getQueriesPeoples().observe(getViewLifecycleOwner(), new Observer<List<Cast>>() {
+
+        mViewModel.getQueriedMovies().observe(getViewLifecycleOwner(), new Observer<List<Movie>>() {
             @Override
-            public void onChanged(List<Cast> people) {
-                mRecyclerViewAdapter.addItems(people);
+            public void onChanged(List<Movie> movies) {
+                mAdapter.addItems(movies);
             }
         });
     }
@@ -68,9 +71,8 @@ public class SearchActorFragment extends Fragment implements CustomizableFragmen
 
     @Override
     public void changeLayout(RecyclerView.LayoutManager layoutManager) {
-      /*  mBinding.recyclerView.setLayoutManager(layoutManager);
+       /* mBinding.recyclerView.setLayoutManager(layoutManager);
         mRecyclerViewAdapter.notifyItemRangeChanged(0, mRecyclerViewAdapter.getItemCount());*/
-
     }
 
     @Override
@@ -78,7 +80,7 @@ public class SearchActorFragment extends Fragment implements CustomizableFragmen
 
         try {
 
-            mViewModel.getPeoplesBySearch(query);
+            mViewModel.getQueriedMovies(query);
         } catch (NullPointerException exception) {
 
         }
