+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ui.ProfileFragment">
+    tools:context=".view.ui.profile.ProfileFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
